refactor: Correct 'loop induction expression should not have float type'
- Change the type of argument 'width' and 'height' to int
- Change loop induction expression to integer type

+++ Includes/CubbyCity/Geometry/Geometry.hpp
@@ -34,8 +34,8 @@ class Geometry
     std::tuple<int, int> ExtractTileRange(const std::string& range) const;
 
     static void BuildPlane(std::vector<PolygonVertex>& outVertices,
-                           std::vector<unsigned int>& outIndices, float width,
-                           float height, int nw, int nh, bool flip = false);
+                           std::vector<unsigned int>& outIndices, int width,
+                           int height, int nw, int nh, bool flip = false);
 
     std::vector<Tile> m_tiles;
     std::vector<std::unique_ptr<PolygonMesh>> meshes;
